requirements.txt
boto3
moto
pytest
mysql-connector-python
faker